# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List all users."
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a user."
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find one user by id or email."
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a user."
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a user."
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/vendors": {
      "get": {
        "operationId": "VendorsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List vendors."
          }
        },
        "tags": [
          "vendors"
        ]
      },
      "post": {
        "operationId": "VendorsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create vendor."
          }
        },
        "tags": [
          "vendors"
        ]
      }
    },
    "/vendors/{id}": {
      "get": {
        "operationId": "VendorsController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find vendor."
          }
        },
        "tags": [
          "vendors"
        ]
      },
      "patch": {
        "operationId": "VendorsController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVendorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update vendor."
          }
        },
        "tags": [
          "vendors"
        ]
      },
      "delete": {
        "operationId": "VendorsController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role-based access header. Pass one of: Super User, Traveler, Travel Partner, Tour Guide, Vendor, Support Executive.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Super User",
                "Traveler",
                "Travel Partner",
                "Tour Guide",
                "Vendor",
                "Support Executive"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete vendor."
          }
        },
        "tags": [
          "vendors"
        ]
      }
    },
...
    "/upload": {
      "post": {
        "operationId": "UploadController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully"
          }
        },
        "summary": "Upload document, image or file attachment (Multipart form-data)",
        "tags": [
          "File Upload"
        ]
      }
    },
    "/upload/{filename}": {
      "get": {
        "operationId": "UploadController_getFile",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Retrieve uploaded file by filename",
        "tags": [
          "File Upload"
        ]
      }
    }
  },
  "info": {
    "title": "Dream Destination Review 4 API",
    "description": "NestJS in-memory REST API with RBAC header authorization for traveler, travel partner, tour guide, vendor, support, and super user portals.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required on every API. Use Super User, Traveler, Travel Partner, Tour Guide, Vendor, or Support Executive."
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "New User"
          },
          "email": {
            "type": "string",
            "example": "newuser@example.com"
          },
          "phone": {
            "type": "string",
            "example": "9999999999"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "role": {
            "type": "string",
            "example": "Traveler"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "Inactive",
              "Suspended"
            ]
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "New User"
          },
          "email": {
            "type": "string",
            "example": "newuser@example.com"
          },
          "phone": {
            "type": "string",
            "example": "9999999999"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "role": {
            "type": "string",
            "example": "Traveler"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "Inactive",
              "Suspended"
            ]
          }
        }
      },
...
      "CreateVendorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ABC Travels"
          },
          "type": {
            "type": "string",
            "example": "Transport"
          },
          "location": {
            "type": "string",
            "example": "Goa"
          },
          "rating": {
            "type": "number",
            "example": 4.5
          }
        },
        "required": [
          "name",
          "type",
          "location"
        ]
      },
      "UpdateVendorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ABC Travels"
          },
          "type": {
            "type": "string",
            "example": "Transport"
          },
          "location": {
            "type": "string",
            "example": "Goa"
          },
          "rating": {
            "type": "number",
            "example": 4.5
          }
        }
      },
...
    }
  }
}